Extract nonce tag transformer into separate file
Also use set to track generated ids to prevent collisions
